Add final css changes
Add final css to move around related links.  While not
complete, this will be the final commit to this project, as
I'm running out of patience tweaking the css to make it
'perfect'.  Instead I plan to move on to another page
design project to 'refresh' my brain, and my motivation.